% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_functions.R
\name{A2}
\alias{A2}
\title{A2 constant}
\usage{
A2(n)
}
\arguments{
\item{n}{Number of samples in the current batch}
}
\value{
Numeric bias correction
}
\description{
A2 constant
}

