% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_functions.R
\name{c4}
\alias{c4}
\title{Bias correction c4}
\usage{
c4(n)
}
\arguments{
\item{n}{Number of samples in the current batch}
}
\value{
Numeric bias correction
}
\description{
Bias correction c4
}

