% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{check_n}
\alias{check_n}
\title{Check n}
\usage{
check_n(n)
}
\arguments{
\item{n}{the parameter n given to the main function constant. Can be an integer of a vector of integers}
}
\value{
Void
}
\description{
This function checks if the parameter n of the main function is within the admissible range
}

