% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{constant}
\alias{constant}
\title{Constants for Shewhart quality control charts}
\usage{
constant(n, cname = c("", "c4", "d2", "d3", "D3", "D4"))
}
\arguments{
\item{n}{Numerosity of sample, must be an integer (numeric) within the [2,25] range.}

\item{cname}{Name of the function to be used, can also be of character class. Admissible values are c4, d2, d3, D3 and D4.}
}
\value{
Unbiasing constant (numeric).
}
\description{
Calculates different constants using a single function
}
\examples{
constant(2, D3)
constant(2,'d2')
constant(2:10, d2)

}

