% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_functions.R
\name{d2}
\alias{d2}
\title{Bias correction d2}
\usage{
d2(n)
}
\arguments{
\item{n}{Number of samples in the current batch}
}
\value{
Numeric bias correction
}
\description{
Bias correction d2
}

