% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_functions.R
\name{d2_}
\alias{d2_}
\title{d2 constant}
\usage{
d2_(n)
}
\arguments{
\item{n}{Number of samples in the current batch}
}
\value{
Numeric
}
\description{
d2 is the mean of the distribution of the relative range.
}

