% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_functions.R
\name{d3}
\alias{d3}
\title{Bias correction d3}
\usage{
d3(n)
}
\arguments{
\item{n}{Number of samples in the current batch}
}
\value{
Numeric bias correction
}
\description{
Bias correction d3
}

