% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_functions.R
\name{D4}
\alias{D4}
\title{Bias correction D4}
\usage{
D4(n)
}
\arguments{
\item{n}{Number of samples in the current batch}
}
\value{
Numeric bias correction
}
\description{
Bias correction D4
}

