% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{f1}
\alias{f1}
\title{Auxiliary function f1}
\usage{
f1(n)
}
\arguments{
\item{n}{Number of samples in the current batch}
}
\value{
Function to be integrated in the d2 function
}
\description{
Auxiliary function f1
}

