% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_functions.R
\name{i_chart_const}
\alias{i_chart_const}
\title{i_chart_const}
\usage{
i_chart_const(n)
}
\arguments{
\item{n}{Number of samples in the current batch}
}
\value{
Numeric constant
}
\description{
This function returns 3/d2(n).
}

