% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_functions.R
\name{one}
\alias{one}
\title{One}
\usage{
one(n)
}
\arguments{
\item{n}{Number of samples in the current batch}
}
\value{
Numeric constant
}
\description{
This function simply returns 1 or a vector of length n with all values set to 1.
}

