% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_functions.R
\name{d3}
\alias{d3}
\title{d3 constant}
\usage{
d3(n)
}
\arguments{
\item{n}{Number of samples in the current batch}
}
\value{
Numeric
}
\description{
d3 is the standard deviation of the distribution of the relative range.
}

