% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_functions.R
\name{D3_}
\alias{D3_}
\title{D3 constant}
\usage{
D3_(n)
}
\arguments{
\item{n}{Number of samples in the current batch}
}
\value{
Numeric constant
}
\description{
Note: an underscore is added for portability. Some machines may not make a distinction between 'd3' and 'D3'
}

